change field type current_weight from integer to float

+++ database/migrations/2016_05_27_195732_add_current_weight_field_to_variants.php
@@ -13,7 +13,7 @@ class AddCurrentWeightFieldToVariants extends Migration
     public function up()
     {
         \Schema::table('variants', function (Blueprint $table) {
-    		$table->integer('current_weight');
+    		$table->float('current_weight');
     	});
     }
